backend/pyproject.toml
[project]
name = "afterglow-api"
version = "1.13.0"
description = "Afterglow - OpenStack VM deployment platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.125.0",
    "uvicorn[standard]==0.30.6",
    "openstacksdk==3.3.0",
    "python-keystoneclient==5.4.0",
    "jinja2==3.1.6",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "python-multipart==0.0.22",
    "httpx==0.27.2",
    "aiofiles==24.1.0",
    "redis[hiredis]==5.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncmy>=0.2.9",
    "prometheus-client>=0.20.0",
    "slowapi>=0.1.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.7.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py312"
extend-exclude = [".venv", "build", "dist", "__pycache__"]

[tool.ruff.lint]
# E/F: pycodestyle/pyflakes, I: isort, B: bugbear, UP: pyupgrade, SIM: simplify
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = [
    "E501",   # line-length 은 format 에서 처리
    "B008",   # FastAPI Depends() 는 default 인자에 함수 호출 필수
    "B904",   # FastAPI except 블록에서 raise from err/None 불필요 (HTTPException 패턴)
    "SIM103", # return 조건 inline 강요 — 가독성 저하 가능
    "SIM105", # try-except-pass → contextlib.suppress — 기존 패턴 유지
    "SIM108", # ternary 강요는 한국어 주석 많은 코드에서 가독성 저하
    "UP042",  # str+Enum → StrEnum — Python 3.11+ 한정, 호환성 위해 유지
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["B", "SIM", "E402"]
# 조건부 서비스 import 패턴 (if settings.xxx: from module import Y) 사용하는 파일
"app/api/container/clusters.py" = ["E402"]
"app/api/identity/admin.py" = ["E402"]
"app/api/network/networks.py" = ["E402"]
"app/main.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.uv]
package = false
